backend/requirements.txt
fastapi==0.112.0
uvicorn==0.34.0
gTTS==2.3.2
openai==0.28.0
python-multipart==0.0.20
pydantic==2.11.4
python-dotenv==1.0.0